Move ForkSpec into World
This patch adds ForkSpec to World (giving `World.ts` control of the `ForkSpec` interface, in the process). This allows us to check in context set-up and code what base we're on, generally for debugging or bootstrapping purposes.
