+++ scenario/CometScenario.ts
@@ -3,7 +3,7 @@ import { expect } from 'chai';
 import { BigNumber } from 'ethers';
 import { World } from '../plugins/scenario';
 
-scenario('initializes governor correctly', {}, async ({ comet, actors }) => {
+scenario('initializes governor correctly', {}, async ({ comet, actors }, world) => {
   // TODO: Make this more interesting, plus the admin here isn't right for mainnet, etc.
   expect(await comet.governor()).to.equal(actors['admin']!.address);
 });
